Cambio de float a double

+++ lab1/funciones.h
@@ -3,8 +3,27 @@
 #include <math.h>
 #include "particula.h"
 
+// Entradas:
+// Salidas:
+// Descripcion:
 Particulas *Lectura_Particulas(char *NameFile);
-float FormulaYComparacion(int celdas, float energia_Inicial, int posicion_Celda, float energia_Particula, int posicion_Particula);
-void MostrarCeldas(float *array_Celdas, int cantidad_Celdas);
-void EscrituraArchivo(float * array_Celdas, int cantidad_Celdas, char * nombreArchivo);
-void MaxEnergiaCelda (float * array_Celdas, int cantidad_Celdas, float *MaxValor, int * posicionMaxValor);
+
+// Entradas:
+// Salidas:
+// Descripcion:
+double FormulaYComparacion(int celdas, double energia_Inicial, int posicion_Celda, double energia_Particula, int posicion_Particula);
+
+// Entradas:
+// Salidas:
+// Descripcion:
+void MostrarCeldas(double *array_Celdas, int cantidad_Celdas);
+
+// Entradas:
+// Salidas:
+// Descripcion:
+void EscrituraArchivo(double *array_Celdas, int cantidad_Celdas, char *nombreArchivo);
+
+// Entradas:
+// Salidas:
+// Descripcion:
+void MaxEnergiaCelda(double *array_Celdas, int cantidad_Celdas, double *MaxValor, int *posicionMaxValor);
